html/semantics/embedded-content-0/the-video-element/video_audioTracks_return_type.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Video Test: audioTracks method of video element</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/TR/html5/embedded-content-0.html#htmlmediaelement">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<video id="video1">
  Your browser does not support HTML5 video.
</video>
<script>

test(function() {
  var v = document.getElementById('video1');
  assert_equals(Object.prototype.toString.call(v.audioTracks), "[object AudioTrackList]", "the return type of audioTracks attribute");
}, "Check if the audioTracks attribute of a video element must return AudioTrackList object");

</script>
